pyproject.toml
[project]
name = "fa-dae2-capstone-tran-quoc-dat-nguyen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faker>=37.6.0",
    "flask>=3.1.2",
    "kagglehub>=0.3.13",
    "openaq>=0.4.0",
    "polars>=1.33.0",
    "psycopg[binary]>=3.2.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "stripe>=12.5.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
